test.amqtt.io dashboard cleanup (#278)
* update the configuration for the broker running at test.amqtt.io to latest format

* removing all of the MUI components from their example dashboard that are no longer needed

* separating out main grid layout into separate components and functions for readability

* updating count up display to handle bytes, kilobytes, etc
