--- docs_test/src/dashboard/Title.tsx.preview
@@ -1,3 +1,0 @@
-<Typography component="h2" variant="h6" color="primary" gutterBottom>
-  {props.children}
-</Typography>
