--- docs_test/src/dashboard/theme/customizations/index.ts
@@ -1,2 +1,0 @@
-export { chartsCustomizations } from './charts';
-export { treeViewCustomizations } from './treeView';
