+++ docs_test/test.amqtt.local.yaml
@@ -1,0 +1,15 @@
+---
+listeners:
+  default:
+    type: tcp
+    bind: 0.0.0.0:1883
+  std-ws:
+    type: ws
+    bind: 0.0.0.0:8080
+plugins:
+  amqtt.plugins.logging_amqtt.EventLoggerPlugin:
+  amqtt.plugins.logging_amqtt.PacketLoggerPlugin:
+  amqtt.plugins.authentication.AnonymousAuthPlugin:
+    allow_anonymous: true
+  amqtt.plugins.sys.broker.BrokerSysPlugin:
+    sys_interval: 2
